api/requirements.txt
astroid==2.11.6
attrs==21.4.0
black==22.1.0
cachelib==0.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dill==0.3.5.1
flake8==4.0.1
Flask==2.2.2
Flask-Session==0.4.0
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.942
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pylint==2.14.2
pyparsing==3.0.7
pytest==7.1.0
python-dotenv==0.19.2
requests==2.27.1
tomli==2.0.1
tomlkit==0.11.0
types-requests==2.27.12
types-urllib3==1.26.11
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.7.0
